Fix recp. completion popup in newer distos
Somehow the hide and show here caused problems with KWin
or some other component in Debian Stretch.
It caused the completionBox to be hidden behind the
MessageComposer.

Hiding and then delaying the show by 500ms in moveCompletionBottom
seems to work nicely though. An immediate single Shot did not work.

A don't ask me why but it works fix.