Add simple key locate to have WKD support
Starts a fire and forget locate-key through GPGME
if the new option to automatically search for
encryption keys is set.

A needed improvement would be to toggle the encrypt
action based on the locate results and respect origin
when calculating trust.

In difference to the first commit this is now guarded
with a gpgme version check to avoid unexpected behavior
with older versions.